Minor fixes
Allows for poweroff command to be called from qml.

This is a first step, and a dirty one. Proof of concept shows
contacting of v2x script on host, passing a command to power off the
unit. This is not an ideal approach. It would instead be better to
simply contact the v2x control serial port directly.

Please note, that for the current smart cities host, the service files
that use the script to contact the port (keep alive) should be set to a
much less aggressive interval, to avoid conflict.